+++ src/PluginManager/src/SkyConnectManager.cpp
@@ -162,10 +162,10 @@ bool SkyConnectManager::freezeUserAircraft(bool enable) noexcept
     return skyConnect ? skyConnect->get().freezeUserAircraft(enable) : false;
 }
 
-bool SkyConnectManager::sendSimulationEvent(SkyConnectIntf::SimulationEvent event) noexcept
+bool SkyConnectManager::sendSimulationEvent(SkyConnectIntf::SimulationEvent event, int arg1) noexcept
 {
     std::optional<std::reference_wrapper<SkyConnectIntf>> skyConnect = SkyConnectManager::getInstance().getCurrentSkyConnect();
-    return skyConnect ? skyConnect->get().sendSimulationEvent(event) : false;
+    return skyConnect ? skyConnect->get().sendSimulationEvent(event, arg1) : false;
 }
 
 SkyConnectIntf::ReplayMode SkyConnectManager::getReplayMode() const noexcept