+++ src/Plugins/Connect/MSFSSimConnectPlugin/src/MSFSSimConnectPlugin.cpp
@@ -149,7 +149,7 @@ bool MSFSSimConnectPlugin::onFreezeUserAircraft(bool enable) const noexcept
     return ok;
 }
 
-bool MSFSSimConnectPlugin::onSimulationEvent(SimulationEvent event) const noexcept
+bool MSFSSimConnectPlugin::onSimulationEvent(SimulationEvent event, int arg1) const noexcept
 {
     HRESULT result {S_OK};
 
@@ -159,6 +159,9 @@ bool MSFSSimConnectPlugin::onSimulationEvent(SimulationEvent event) const noexce
         break;
     case SimulationEvent::EngineStop:
         result = ::SimConnect_TransmitClientEvent(d->simConnectHandle, ::SIMCONNECT_OBJECT_ID_USER, Enum::underly(SimConnectEvent::Event::EngineAutoShutdown), 0, ::SIMCONNECT_GROUP_PRIORITY_HIGHEST, ::SIMCONNECT_EVENT_FLAG_GROUPID_IS_PRIORITY);
+        break;
+    case SimulationEvent::SimulationRate:
+        result = ::SimConnect_TransmitClientEvent(d->simConnectHandle, ::SIMCONNECT_OBJECT_ID_USER, Enum::underly(SimConnectEvent::Event::EngineAutoShutdown), arg1, ::SIMCONNECT_GROUP_PRIORITY_HIGHEST, ::SIMCONNECT_EVENT_FLAG_GROUPID_IS_PRIORITY);
         break;
     case SimulationEvent::None:
         // Nothing to do